refactor: vite as builder, refactor almost everything
