+++ .env.example
@@ -1,0 +1,2 @@
+NODE_ENV=test
+API_URL=""
