--- craco.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-    style: {
-        postcss: {
-            plugins: [
-                require('tailwindcss'),
-                require('autoprefixer'),
-            ],
-        },
-    },
-}
