+++ package.json
@@ -1,16 +1,13 @@
 {
   "name": "wargalokal-website",
   "version": "0.1.0",
-  "private": true,
   "dependencies": {
-    "@craco/craco": "^6.2.0",
     "@fortawesome/fontawesome-free": "^5.15.4",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^26.0.24",
     "@types/node": "^12.20.19",
-    "@types/react-dom": "^17.0.9",
     "axios": "^0.24.0",
     "chart.js": "^3.5.1",
     "react": "^17.0.2",
@@ -19,14 +16,12 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
-    "typescript": "^4.3.5",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
-    "start": "craco start",
-    "build": "craco build",
-    "test": "craco test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
   },
   "eslintConfig": {
     "extends": [
@@ -47,10 +42,16 @@
     ]
   },
   "devDependencies": {
+    "@types/react": "^17.0.38",
     "@types/react-alert": "^5.2.1",
+    "@types/react-dom": "^17.0.11",
     "@types/react-router-dom": "^5.1.8",
-    "autoprefixer": "^9.8.6",
-    "postcss": "^7.0.36",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.7"
+    "@vitejs/plugin-react": "^1.0.7",
+    "autoprefixer": "^10.4.2",
+    "parcel": "^2.2.1",
+    "postcss": "^8.4.5",
+    "tailwindcss": "^3.0.17",
+    "typescript": "^4.4.4",
+    "vite": "^2.7.2"
   }
 }
