+++ src/api/village/Village.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
-import { apiUrl } from "src/shared/env";
-import { combineURLPaths } from "src/utils/urlHelper";
+import { apiUrl } from "../../shared/env";
+import { combineURLPaths } from "../../utils/urlHelper";
 
 export const getAllUser = async () => {
   const response = await axios.get(combineURLPaths(apiUrl, `api/v1/user/get-all`));
