+++ src/styles/globals.css
@@ -1,8 +1,8 @@
+@import url("https://fonts.googleapis.com/css2?family=Nunito&display=swap");
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-@import url("https://fonts.googleapis.com/css2?family=Nunito&display=swap");
 
 * {
   font-family: "Nunito", sans-serif;
